refactor: one module per document type, collapse assess logic into type modules
ResearchDocument + assess/gate logic → wos/research.py
PlanDocument + _parse_tasks + assess logic → wos/plan.py
ContextDocument → wos/context.py
document.py → Document base class only, lazy imports for all subclasses
Delete wos/assess_research.py and wos/assess_plan.py
