+++ scripts/_bootstrap.py
@@ -1,0 +1,27 @@
+"""Bootstrap helper: insert the plugin root into sys.path.
+
+Import as a side effect before importing from wos:
+
+    import _bootstrap  # noqa: F401
+
+When run as ``python scripts/<name>.py``, Python adds ``scripts/`` to
+sys.path[0], so ``import _bootstrap`` finds this file. It resolves the
+plugin root and inserts it so that ``import wos`` succeeds.
+
+The resolved root is also exposed as ``plugin_root`` for scripts that
+need a reference to it at runtime.
+"""
+from __future__ import annotations
+
+import os
+import sys
+from pathlib import Path
+
+_env_root = os.environ.get("CLAUDE_PLUGIN_ROOT", "")
+# scripts/ → plugin root
+plugin_root: Path = (
+    Path(_env_root) if _env_root and os.path.isdir(_env_root)
+    else Path(__file__).resolve().parent.parent
+)
+if str(plugin_root) not in sys.path:
+    sys.path.insert(0, str(plugin_root))
