+++ scripts/reindex.py
@@ -11,21 +11,10 @@ Usage:
 from __future__ import annotations
 
 import argparse
-import os
 import sys
 from pathlib import Path
 
-# Ensure `import wos` works whether pip-installed or run from plugin cache.
-# Prefer CLAUDE_PLUGIN_ROOT env var (set by Claude Code for hooks/MCP);
-# fall back to navigating from __file__ (required for skill-invoked scripts).
-_env_root = os.environ.get("CLAUDE_PLUGIN_ROOT", "")
-# scripts/ → plugin root
-_plugin_root = (
-    Path(_env_root) if _env_root and os.path.isdir(_env_root)
-    else Path(__file__).resolve().parent.parent
-)
-if str(_plugin_root) not in sys.path:
-    sys.path.insert(0, str(_plugin_root))
+import _bootstrap  # noqa: F401 — side effect: adds plugin root to sys.path
 
 
 def main() -> None:
