+++ wos/discovery.py
@@ -1,14 +1,13 @@
 """Document discovery via project tree walking.
 
-Walks the project tree from root, respects .gitignore patterns,
-and identifies WOS-managed documents by frontmatter presence.
+Walks the project tree from root, skipping hidden directories and known
+build/tooling directories, and identifies WOS-managed documents by
+frontmatter presence.
 """
 
 from __future__ import annotations
 
-import fnmatch
 import os
-from dataclasses import dataclass
 from pathlib import Path
 from typing import List, Optional
 
@@ -19,130 +18,17 @@ from wos.document import Document, parse_document
 # Pass exclude_dirs=set() to discover_document_dirs() to include all.
 INDEX_EXCLUDED_DIRS = frozenset({".agents", "scripts", "skills", "tests"})
 
-# ── Gitignore parsing ─────────────────────────────────────────
+# Directories always skipped during discovery — hidden dirs (starting with ".")
+# are handled separately in _should_skip_dir().
+_SKIP_DIRS = frozenset({
+    "node_modules", "__pycache__", "venv", ".venv",
+    "dist", "build", ".tox", ".mypy_cache", ".pytest_cache",
+})
 
 
-@dataclass
-class _GitignorePattern:
-    """A parsed .gitignore pattern."""
-
-    pattern: str
-    negated: bool
-    dir_only: bool
-
-
-def load_gitignore(root: Path) -> List[_GitignorePattern]:
-    """Read root/.gitignore and return parsed patterns.
-
-    Returns empty list if no .gitignore exists.
-
-    Args:
-        root: Project root directory.
-
-    Returns:
-        List of parsed gitignore patterns.
-    """
-    gitignore_path = root / ".gitignore"
-    if not gitignore_path.is_file():
-        return []
-
-    try:
-        text = gitignore_path.read_text(encoding="utf-8")
-    except OSError:
-        return []
-
-    patterns: List[_GitignorePattern] = []
-    for line in text.splitlines():
-        line = line.strip()
-        # Skip blank lines and comments
-        if not line or line.startswith("#"):
-            continue
-
-        negated = False
-        if line.startswith("!"):
-            negated = True
-            line = line[1:]
-
-        dir_only = False
-        if line.endswith("/"):
-            dir_only = True
-            line = line.rstrip("/")
-
-        if not line:
-            continue
-
-        patterns.append(_GitignorePattern(
-            pattern=line,
-            negated=negated,
-            dir_only=dir_only,
-        ))
-
-    return patterns
-
-
-def is_ignored(
-    path: Path,
-    root: Path,
-    patterns: List[_GitignorePattern],
-) -> bool:
-    """Check if a path should be ignored based on gitignore patterns.
-
-    Always ignores ``.git/`` regardless of patterns.
-
-    Args:
-        path: Absolute or relative path to check.
-        root: Project root directory.
-        patterns: Parsed gitignore patterns from load_gitignore().
-
-    Returns:
-        True if the path should be ignored.
-    """
-    try:
-        rel = path.relative_to(root)
-    except ValueError:
-        return False
-
-    rel_str = str(rel)
-    parts = rel.parts
-
-    # .git/ is always ignored
-    if parts and parts[0] == ".git":
-        return True
-
-    is_dir = path.is_dir()
-
-    # Check each path component individually against patterns
-    # (gitignore patterns match at any directory level by default)
-    matched = False
-    for pat in patterns:
-        if pat.dir_only and not is_dir:
-            # Directory-only patterns only match directories, but we also
-            # check if any parent component matches
-            for i, part in enumerate(parts):
-                if fnmatch.fnmatch(part, pat.pattern):
-                    # This component is a directory (all but last are dirs)
-                    if i < len(parts) - 1:
-                        if pat.negated:
-                            matched = False
-                        else:
-                            matched = True
-                        break
-            continue
-
-        # Check if pattern contains a slash (path-based match)
-        if "/" in pat.pattern:
-            # Match against the full relative path
-            if fnmatch.fnmatch(rel_str, pat.pattern):
-                matched = not pat.negated
-        else:
-            # Match against any path component or the filename
-            component_match = any(
-                fnmatch.fnmatch(part, pat.pattern) for part in parts
-            )
-            if component_match:
-                matched = not pat.negated
-
-    return matched
+def _should_skip_dir(name: str) -> bool:
+    """Return True if a directory should be excluded from discovery."""
+    return name in _SKIP_DIRS or name.startswith(".")
 
 
 # ── Document discovery ────────────────────────────────────────
@@ -155,7 +41,8 @@ def discover_documents(root: Path) -> List[Document]:
     (``name`` and ``description`` fields). ``_index.md`` files are
     excluded (they are generated artifacts).
 
-    Respects ``.gitignore`` patterns and always skips ``.git/``.
+    Skips hidden directories (starting with ```.```) and common build /
+    tooling directories (``node_modules``, ``__pycache__``, etc.).
 
     Args:
         root: Project root directory.
@@ -163,30 +50,16 @@ def discover_documents(root: Path) -> List[Document]:
     Returns:
         List of parsed Document instances, sorted by path.
     """
-    patterns = load_gitignore(root)
     documents: List[Document] = []
 
     for dirpath, dirnames, filenames in os.walk(root):
-        dp = Path(dirpath)
-
-        # Filter out ignored directories in-place (prevents os.walk descent)
-        dirnames[:] = [
-            d for d in dirnames
-            if not is_ignored(dp / d, root, patterns)
-        ]
-        dirnames.sort()
+        # Filter out skipped directories in-place (prevents os.walk descent)
+        dirnames[:] = sorted(d for d in dirnames if not _should_skip_dir(d))
 
         for filename in sorted(filenames):
-            if not filename.endswith(".md"):
+            if not filename.endswith(".md") or filename == "_index.md":
                 continue
-            if filename == "_index.md":
-                continue
-
-            file_path = dp / filename
-            if is_ignored(file_path, root, patterns):
-                continue
-
-            doc = _try_parse(file_path, root)
+            doc = _try_parse(Path(dirpath) / filename, root)
             if doc is not None:
                 documents.append(doc)
 
@@ -227,6 +100,37 @@ def discover_document_dirs(
             continue
         dirs.add(d)
     return sorted(dirs)
+
+
+def filter_documents(
+    root: Path,
+    doc_type: str,
+    subdir: str = "",
+    status: str = "",
+) -> tuple[str, List[Document]]:
+    """Discover documents filtered by type, optional status, and optional subdir.
+
+    Args:
+        root: Project root directory.
+        doc_type: Required document type (e.g. 'research', 'plan').
+        subdir: Optional path prefix to restrict results (relative to root).
+        status: Optional status value to filter on (e.g. 'executing').
+
+    Returns:
+        Tuple of (label, docs) where label is the scanned path string and
+        docs is the filtered list of Document instances.
+    """
+    docs = discover_documents(root)
+    docs = [d for d in docs if d.type == doc_type]
+    if status:
+        docs = [d for d in docs if d.status == status]
+    if subdir:
+        docs = [
+            d for d in docs
+            if d.path.startswith(subdir + "/") or d.path.startswith(subdir)
+        ]
+    label = os.path.join(str(root), subdir) if subdir else str(root)
+    return label, docs
 
 
 def _try_parse(file_path: Path, root: Path) -> Optional[Document]:
